Feature discovery for contribution menu
If a user accessed CX using a non-persistent entry point,
show a 'feature discovery' for contribution menu next time
the user visit any wiki page.

This is shown atmost once to user.
The information about whether this is shown is stored in local
storage. Reason: The conditions to show this feature discovery
is narrow and one time occuring. Not having the local storage
value means no difference to user.

Bug: T216498
Change-Id: I79fefdfdb69fe34aabfca39231af329933e0d0bd