Make CX functional in Minerva skin
An invalid assumption about the presence of personal tools was
causing a js error while rendering the (custom) header. This
was resulting blank page.

In this patch, I check for the presence of personal tools and render it.
It it is not present, like the case of Minerva, in the place of
personal menu, just add a button that shows user name and icon.
Clicking on that takes the user to user page.

We can think about making the user menu same as other skins,
but as follow up.

Bug: T208159
Change-Id: Ib85ba92e5d72695ae23b033e1f852eefb86ae326