Remove $wgContentTranslationEventLogging
It provides little utility compared to the amount of code it takes
to support it. Since mw.track is a no-op if EventLogging isn't
available, there is little need to avoid loading the modules. The
one call in PHP is already guarded by extension check. Turning
all event logging off for CX in production would be harmful for
our analytics.

Change-Id: Ie3baf633c127c9b457017c7c4dab65118e5e68e9