Actually enforce valid token for translation view
Previous attempt changed the token validation to check that the
target language matches current wiki. However, it was not sufficient
since the code that determines whether we are on a translation view
was funky: valid token was sufficient, but not necessary requirement.
One could always access existing translations, as long as one was
logged in and owner of the draft. This is no longer the case and a
valid token is now always required.

Bug: T154888
Change-Id: Id1c20d4eb44b35166fc0bd62b802510e2baddd58