Return the last translation while querying for a language pair and title
If there are multiple translation attempts for same language pair and source
title, return the last one. The first one would be mostly uninteresting.

The last one can be used to inspect a published translation for debugging
purpose.

Can be improved to return all translations, but I am not attempting that
here.

Change-Id: Id3f82b83a46eb800eac7cef6b17c1d88dbe5944f