Do not use DB_MASTER for suggestion list reads
This causes logspam in logstash. Some of the callers first  check
if list exists, and if not, proceeds to create it. This introduces
a bigger chance for race conditions (they already existed before).

Manipulating suggestion lists is not a such critical operation so
we can deal with some failures caused by race conditions.

Change-Id: Idc617f281e40774b99528152e5752248389fc7cf