Use quick checks for userCan
This hook is pretty much run on every page view, and it defaults
to "secure" rigor which does queries on master servers. Fast checks
are sufficient for us, because this is only whether we show the UI,
not actually enforcing it. This caused logspam in logstash.

While at it, replace deprecated Title::userCan with PermissionManager.

Change-Id: Ibbc5a9a072f97dee85f033dcdd4fd48c8c4850a9