Use original revision when latest one causes data loss
When restoring an in-progress translation for which there
is a new revision of the source article, do the following:
* If the latest revision of the source can be loaded without causing
any data loss to the user translation, load the new revision.
* If the latest revision cannot be loaded without causing any
data loss to the user translation, load the original revision
and show the warning - T168287.

In order to achieve this behavior, change initialization of
mw.cx.dm.Translation and create source and target DOM
nodes in mw.cx.init.Translation, then pass to constructor.
First, create target DOM in order to check for data loss
(sections not restored). If no sections are lost or new
translation is being started, proceed with creating
rest of objects in same manner as before this change.
If there is any lost section, fetch source page of the
original revision used when starting translation.
Then go through mostly same process as when there's no
data lost, but show warning that content has changed
and update URL.

Bug: T220790
Change-Id: I47fdc93736610459228c16fae90ce6e5cf337f77