Show the invitation to translate when creating new one, more than once
In T216032 we limited the inviation with the following conditions
1. A user who is not a translator yet
2. A multilingual user
3. CX is not a beta feature
4. User has not dismissed the invitation yet.

Here we are removing conditions #1 and #4 from above list and showing
this invitation more than once. But in follow up commit we will add
another condition - that we have a source article suggestion to give.

To remove condition #1, isPotentialTranslator method now returns true
if user is already a translator.

To remove condition #4, we no longer set global preference
cx_campaign_newarticle_hide

Bug: T216500
Change-Id: I03dda57393e4976a716b058202eadf77904ab687