Show new translation dialog when target language and title are present
When CX dashboard is loaded, if there is all information needed
to pick article to translate, we prefill New translation dialog.
Data needed for that is source and target language pair, and
source title. On the other hand, if we have target language and
target title information, we could show New translation dialog
with search bar open, where source language and article can be
found. Examples where that data is provided is when uses sees
information that someone else owns the translation, but then
they click on button there to start a new translation, where
they could pick new source article.
Also, for 'newbytranslation' campaign, user is provided with
target article and language, where we can allow them to pick
source article.

Bug: T229419
Change-Id: I44e468c2775432999dc100a2cb7db19444d7889d