Keep newly favorited articles after changing language pair
Favorite articles (in "for later" list) should be displayed
regardless of language pair selected. However, if new article
is added to the list and source/target language is changed,
memory structure holding suggestions for previous language
pair still has the article which is moved to "For later",
which removes it upon language change action.
To prevent this, remove favorited articles from suggestion
memory structure.

Bug: T194476
Change-Id: If0cc738283b06d4eaa75d28fa7204455d1a3771e