Show Abuse Filter errors on sections even if there is no message body
If "Prevent the user from performing the action in question" for AF rule
is checked, but "Trigger these actions after giving the user a warning" is not,
we get error during section validation, but there is no message HTML for
that error. Our code ignores such errors with no messages, but since this
is preventive error that probably affects user's publishing, we should
display it, even if that means using some generic message title.

Change-Id: Iae6e22e547a876c53bda679c88b7cce6b882e12b