Fix the default result of isPotentialTranslator
If none of the conditions applies, the method was returning true,
indicating that the user is a potential translator. Fix it.

The user is a potential translator only when user has CX translations
or contributions to more than one wikipedia.

Because of this issue, for a new user, the invite was always showing
even though the user has not made any edits in any wiki.

Bug: T237607
Bug: T232372
Change-Id: If84f21654f493f80f0a23858f601c7b0fb424294