Document origin and reason for core code duplication
These references will hopefully discourage local mutations,
within that subsection, and also offer a way to get to the
original in case it needs updating. It'll also allow this code
to be found when grepping for callers of the original when
considering to make changes to it.

This also moves the loadSkinModules() call to match the order
of execution as it would normally be in MediaWiki core.

Bug: T243286
Change-Id: Ic7d1214ac915d4eaaace07f8b93a7aa1e3846d42