Avoid failure on restoring draft with no categories
Apparently, if translator does not touch categories, we are not saving
them in the drafts. Instead of returning an empty array, we get a null.

Since it is possible we would like to keep the distinction between no
categories saved and no categories, I opted to just add a guard before
the code that is supposed to be temporary anyway.

Bug: T249400
Change-Id: Ifb0eec66e7d334ab6733ffaebf37e6a28eae5481