dashboard: Validate the bundle created by webpack
The dist folder in CX dashboard app does not get reviewed as it is
minified bundle created by webpack using vue-cli tooling. Committing
the dist folder to gerrit is not an optimal method, but at present
we do not have choices. To make sure that the dist folder is not
manipulated by any reason, have a CI test that also create a dist
and check for diffs. So that nobody can commit a bundle with injected
code in it.

Secondly, make sure that bundle size(gzipped) is in control using
the bundlesize tooling.

This patch contains the bundle validation script, bundlesize check,
and CI integration of cxdashboard tests

Thanks to Job Robson for this inputs.

Bug: T247499
Change-Id: I66266b6cc40b10010709b502a05ed00c0da97771