MT Abuse: Enforce min length consideration for sections with excluded children
Current implementaton calculates total length for a section by considering both excluded and
non-excluded nodes.That leads to a total length that also counts excluded nodes, which leads
to false positive MT Abuse warnings. This patch introduces a way to calculate total section
length by ignoring excluded children nodes even when parent node has multiple children and
some of these children happen to be non-excluded from MT Abuse validation. The new functions
were made static (as well as isExcludedFromValidation) as it is easier to test and there is
no apparent reason not to be static. Tests were added to verify the functionality.

Bug: T232718
Change-Id: Ie7f7f33d1b8e9c6f1c0d7858e0c9bcbcad64e80f