Refactor: Use model classes
Instead of using generic objects, use well defined model classes.
Define models for Translation, Translator, Article, Suggestion and
SectionSuggestion. This also detaches the whole code from the structure
of api output. API output is transformed to these model classes.

Defining model classes simplified code in many Vue Component classes
and avoided many hacks.

For developers, the data structure is well defined and easy to work with.

Vue properties enforce that the passed value is instance of the above classes.
Previously it was not possible because we just defined the type as Object.

Renamed 'Article' references to 'Page'
And 'Suggestions' to 'PageSuggestions', but there are few more instances left.
It can be done in a simple patch to avoid merge conflicts.

Bug: T241591
Bug: T243494
Bug: T243430
Bug: T247499
Change-Id: I2916c80e5a4f866577dab687dc95871bf1b946f2