Grid in scss and Expose viewports to js
Rewrite the grid in scss for better readability and maintainability.

It also allows exporting style variables to javascript to use
in plugin.

Add a plugin to report the current viewport to javascript. It is
not used anywhere yet. Illustrated it using storybook.

More helper classes in grid are documented in storybook.

Change-Id: Ica497085bbee4705a1b28f03836fa525cd21f533