UI framework: Typography
An attempt to implement Wikimedia typography styleguide
https://design.wikimedia.org/style-guide/visual-style_typography.html

No helper/semantic classes are defined now. We will require that in the future.
Styleguide does not define a design vocabulary for that either. If we use
correct html5 tags instead of generic span, div tags this can be addressed
to certain extend, but not sufficient.

We will need per script font stacks. Current font-stacks support only
latin scripts. It is a non-trivial effort and can be done in iterations.

Change-Id: I99117ee0c24d4502ddf9dc074c56ce629fec18e2