CX language info: create autonym mixin and delete MWAutonym
Several components use languageInfo and autonym method to get the autonym
of a language. That leads to code duplication and once we decide to change
our approach on how to compute autonyms, we 'll have to replace multiple
instances of this code snippet. Thus I created this mixin so that we can use
it everywhere. Additionally, MWAutonym was deleted as it doesn't offer any
additional functionality.

Bug: T241587
Change-Id: Ie285cf6edbfcf67ebf0c7360b1c036bfdb4b97c9