Build: Prevent ResourceLoader minification
Due to T255556, we need to avoid extra minification on build
assets of vue CX Dashboard. Append a comment /*@nomin*/ to
get this effect from ResourceLoader. Do the required configration for this
in vue.config.js

While doing this configuration, avoid code obfuscation/mangle in
minified code. For our opensource project, this is not needed and
does not change the output size much thanks to gzip compression.

Change-Id: Ie1471b4dc78aa3466ff1430dc3f4f2c90b005ae2