SX Compare Contents: Create component skeleton
Create a first version for SXContentComparator component. Lots of
details still to be implemented. Several bugs in this patch are
fixed in subsequent patches in this chain, as the goal of the
current patch is to create a (quite unrefined) component that
will get gradually enhanced with all required functionality
and proper layout.

Bug: T241589
Change-Id: Ie1dd7258ac3edc505af47d166d4616fd543dcec3