CX Vue Dashboard: Fix bottom navigation
Made layout and style changes to fix bottom navigation to match
design specifications. Adaptations were made to MwBottomNavigation
component, so that default layout follows SX design specifications.
This component remains flexible enough for configuration as a slot
is provided to overwrite MwButtonGroup if needed and default styles
only apply to default MwButtonGroup class, that can also be omitted
if these styles are not wanted.

Bug: T255749
Change-Id: I685fd913f1183cc2903bb118148c8e5fb14f66a8