Combine single-use resource loader modules into `ext.cx.tools`
Only used in one place, as a dependency for `ext.cx.tools`; just
add the javascript, styles, etc. to that module directly.

Modules merged:
* ext.cx.tools.dictionary
* ext.cx.tools.formatter
* ext.cx.tools.gallery
* ext.cx.tools.poem
* ext.cx.tools.images
* ext.cx.tools.instructions
* ext.cx.tools.linter
* ext.cx.tools.link
* ext.cx.tools.mt
* ext.cx.tools.mt.card
* ext.cx.tools.mtabuse
* ext.cx.tools.categories
* ext.cx.tools.reference
* ext.cx.tools.template
* ext.cx.tools.template.card
* ext.cx.tools.template.editor

All these modules are part of Content Translation version 1
which is not supported and kept only for continuing translations
that were started with it. In upcoming months we will remove this
code.

Bug: T231326
Change-Id: I4e243f791e0a04d0849a9e97ec7a6753587bdbd7