CX Suggestion api and store module: Fix to handle api errors
This patch fixes the following issue: When a user requests to
enhance a translation (with SX) for a language in which the
article doesn't exist, section suggestion request fails with
an error and corresponding fetchPage request is never being
sent. Although this patch isn't expected to insert any issues,
no exhaustive investigation has been made to check if anything
is affected. An error handling service can also be developed
in future iterations.

Bug: T241590
Change-Id: I6d839d2aedc87e44032a93b326641f382c98575e