MWButton: Always add padding for icon and indicator unless icon type
Padding is currently being added to icon/indicator only when they
go before/after a button label. This means that no padding is being
added when button is created using custom content (overriding the
slot). Since the only case where padding is undesirable is when
button is of type icon, css rules were changed to reflect that.

Change-Id: Ia580a4b7726ed16ec38fd7c97bbce600aae16e1b