SX Language info: Fetch from a production wiki instead of local wiki
Fetching the language info from development or test machines were taking
more time and affected by slow apis. This api depends on localization cache
and if it is not properly populated, can take 100s of requests to fullfill
the request. In production wikis, the api is fast and does not need multiple
queries for pagination. The language information is same no matter which wiki
we query.

Change-Id: Ib0f828272e1600908164aff5d8d6251d15f3acde