CXDashboard: Fix infinite spinner for empty translation lists
Currently when no translation exists for a translation list, loading
spinner is being shown infinitely. This patch adds a new state
variable named "translationsLoaded", which is set to true when
fetchTranslations action is finished (even when returning an empty
translations array). That way spinner is only shown when
fetchTranslations action is still pending.

Bug: T253253
Change-Id: I5e5d8cfd58461619325d8cc733842c9587323eb1