SXContentComparator: Fix content on current section change
Currently when user navigates to another section content inside
content comparator remains the same. This is ok when new section
has the same mapped status, but leads to inconsistent UI when not.

Example: When current section is a missing section, available
content options are "Section in source language" and "Full
article in target language". If user selects "Full article"
content and then navigates to the next section, which is a
present section, content options are updated to "Section in
source language" and "Section in target language", but content
still contains full article, which is confusing for the user.

Solution: Watch for isMappedSection prop and update content
accordingly to updated prop values.

Change-Id: I1320805dbd259849084c8bb648555c022d639255