SX sectionSuggestion model: Add section order information
This patch adds "sourceSections" and "targetSections"
properties to SectionSuggestion model, which store
information about section order inside source and target
page respectively. The information was already existing
in the suggestion api response but not used until now.
A method that returns the section order number for a given
section title is also added

Change-Id: I59fdef8cf2f843c42eb6c6367d2119ba9d541fe8