Section Translation: Introduce subsection model
Section content is currently served to the user by
using section sentences. That makes content being
displayed in a weird way: https://imgur.com/a/xuCJwSN

To fix that, this patch introduces subSection model
which is used for proper content rendering. This model
enables the integration of all types of content,
images etc as can be seen here:
https://imgur.com/a/B75QwLC

Change-Id: I0aa6043232fe708053418ae8517340bf7e54278a