SX: Group MW page style customization, clear floating figures
In CX1 and CX2, we avoided floating image since it does not make
sense to have it in a narrow content space. Apply the same style
in SX too.

Also group the page style related customization to a single LESS
file.

Remove the line height 1.3 to pick that from typography.scss. 1.3 is
too small for non-latin scripts.

Change-Id: If857a9ce5eddea94eb258e72869f447b9082e303