Fix SX Editor redirection on editor close
When user closes (either by pressing the close or
the next button) the sx editor, they are redirected
to "Pick a sentence" screen. However, sx editor is
always being used inside "Preview and Publish" step.
Fix this by adding a navigation guard and pass the
previous route as an argument to SXEditor, so that
redirection leads to previous step.

Issue screencast: https://imgur.com/a/sfKQZoS

Change-Id: I29c7e01c93c82c3cb3b109250d160769fd917f1c