CXSuggestionList: Start with stored sectionSuggestion if exists
Currently when CXSuggestionList SFC is mounted, it checks for URL
params and start a new translation with a dummy sectionSuggestion
(until the proper sectionSuggestion is loaded from api in the
background).

When user revisits CXDashboard, this is happening again, even though
this sectionSuggestion has already been loaded and is present inside
Vuex store.

Fix this by checking for existing sectionSuggestion before creating
a dummy one, and use this instead if exists.

Change-Id: Ic97203b9f57adfe038a3e1646a729a6cd34e65b0