SX Refactoring: Rename proposed translation card action buttons SFC
Component implementing the action buttons for proposed translation
card, is wrongfully named SXSentenceSelectorActionButtons. This can
be confusing as these buttons are specific to untranslated sentences.
In case of an already translated sentence, the corresponding card is
different and will be implemented in different components.

Fix this by changing the name to a more meaningful one.

Change-Id: I773db2032f18773eca9465c3ddcda6c4d816699c