SX Article Selector refactoring: Create existing article body
In continuation of the previous patch in chain, this patch introduces
ExistingArticleBody SFC, in which the content for the case of the
already existing article is encapsulated.

While at it, a major refactoring is made to:

- Remove unnecessary prop binding and use vuex store instead
- Use arrow syntax for computed properties whenever possible
- Also refactor SXArticleSelectorHeader following the same logic

Once, this refactoring is done, we can easily integrate another SFC
like ExistingArticleBody, named "PrefilledSectionBody", that
encapsulates the content for the case of direct access to a prefilled
section. (example name: PrefilledSectionBody)

Change-Id: Iec2fa986948a0056c77cbd7f0f2c0a6ea72409ad