Pick a sentence: Fix translation edit result for empty text provider
Currently when user selects EMPTY_TEXT provider, then edits the
translation and saves, they cannot any longer select this translated
sentence. This is an issue that existed before for section title,
too, and it is caused because VE adds editor content in a new
paragraph, when empty content is provided on VE load.

Fix this by providing an empty dummy span node to VE, instead of
nothing, when proposed translation is empty, and remove it when
editing is completed.

Bug: T269926
Change-Id: Iff392b4be8f0250256c24c1d5af941487c007dd2