SX Router: Introduce force param on routes
Currently, a navigation guard exists that validates navigation
between different steps in the application, based on their expected
visiting order. This exists so that we can ensure the appropriate
content is always available to the user and that they cannot visit
last steps without vising the intermediate workflow steps.

However, there are some use cases that we want to bypass this
validation.

To support that need, introduce force param on routes so that
validation can be bypassed when this param is set to true.

Change-Id: I67a3666e5a714867eb0fcfa0e10523e7f432a029