CXDashboard: Move all elements from Dashboard view to CXDashboard SFC
Currently, Dashboard view component contains several elements. This
is inconsistent with the rest of view components that only hold the
corresponding screen root component.

To fix this, move all the elements inside CXDashboard root component.
While at it, refactor some minor layout issues (not needed rows, cols
etc) and adjust UI elements properly.

This patch is a prerequisite for the addition of the
ExperimentalSupportBanner SFC that will be added in subsequent patch.

Change-Id: I8de2c368e34144d6de2aeea4352f43bbe1777e9e