SX Highlight published section after publishing and scroll to it
According to our design, upon successful SX publishing, user should
be redirected to target article page, and specifically to the newly.
published section. Additionally, the section should be highlighted
and a "new section" indicator should be added to it.

Design: https://phab.wmfusercontent.org/file/data/3zhjufaqcxzccevk43t5/PHID-FILE-nshl4tmgmvpfjgrneqbl/Contribute_again_-_Highlight.png

To support this request, a new "addSXPublishingFollowupModule"
listener to "BeforePageDisplay" hook is added, that adds
"sx.publishing.followup" module for all content pages. This module
checks if "published-section" URL param exists and is populated with
a valid section title. If such a section is found, it is highlighted.

Bug: T260139
Change-Id: I1521965aa30aec61749849f38e6780dbb7fb26a4