CX-SX Switch: Enable CX on desktop and SX on mobile devices
As part of experimental SX support for mobile devices, CX modules are
now updated to only target desktop devices, while SX modules (mw.cx3)
are targeting exclusively mobile devices.

"ContentTranslationVueDashboard" configuration variable is currently
renamed to "ContentTranslationEnableSectionTranslation", and is being
used as a switch to control if section translation application will
be rendered on mobile screens or not at all. If set to false, Content
Translation will be rendered for both desktop and mobile devices.

Bug: T268430
Change-Id: I51414f33bfd5f912f7905fa761b1f5c36531f79a