SX Dashboard: Fix suggestion fetching when no published translations
Currently, user published translations are being used as seed for
section suggestion fetching. This way when a user has no published
translations (which is the case for new users that we are targeting),
no sections suggestion are being displayed to the user.

Fix this by using cx published translations for the selected language
pair as seed for section suggestions, when no user published
translations exist.

Section suggestion are being fetched and stored one by one, so that
user sees results as soon as possible.

Change-Id: I78db3ae9b28efcc7348086bb5a6b014063262fa8
Bug: T272542