SX Publisher: Deactivate translationInProgress before redirection
Currently translationInProgressExists is a vuex getter that checks if
section title or sentence has been translated. Instead of using a
getter, it would be easier to store this value in a state variable
and modify it appropriately.

This way we can easily deactivate warning before leaving page upon
successful publishing.

Expected functionality: warning is shown when user has translated
at least one segment (title or sentence) but not when page is
redirecting after publishing.

Change-Id: Ic1b03024d05e82c3d2c3880f0e2530060f2f665f