SX Rename "proposed translation" to "suggested translation"
We want to update the title for the proposed translation card,
so that we use "suggestion" concept for consistency.

This patch updates translation message as needed and also renames it,
to follow that concept. However, related components and variable
names in our codebase are not renamed, to avoid unnecessary overhead.

Bug: T273053
Change-Id: I3d336111fc7ea3d4f9117acc870c7b0f2a3e780b