SX: Add section translation target language configuration parameter
Since we are moving towards deployment on Bengali wiki, we need to
make sure only this language is supported, in order to avoid broken
functionality or user confusion.

Fix this: by adding SectionTranslationTargetLanguage configuration
parameter inside extension.json. When set, this configuration
parameter is used as the only available target language within the
application.

While at it SXArticleLanguageSelector is refactored to use
SXTranslationListLanguageSelector SFC as their functionality is
identical.

Bug: T273274
Change-Id: Iad06d9abca1eae9bb1e0e94f56c429a0941abf02