SX Fix TranslatedSegmentCard for empty translation MT option
Currently TranslatedSegmentCard is not appearing for translations
started with empty sentence MT option. This is caused by the fact
that MT score is calculated with a division that uses proposedText
length as denominator. As easily understood, when this text is equal
to zero an exception is being thrown, leading TranslatedSegmentCard
to disappear.

Fix this by return 100% for MT score when proposed translation length
equals to zero.

While at it use Math.ceil for MT score calculation so that even the
slightest change is being detected so that we can provide better
feedback to user.

Bug: T273444
Change-Id: I21c9579488b9bf282a28bd19e2efea5f7ceb4ece