CX2: Restore scroll positions on machine translations
It was observed that when the section placeholder is clicked on
an article translation with some already translated section, the scroll
position jumps. This annoying behavior was mostly seen in Firefox
in my testing.

The changing of MT provider was free from this bug and it was found that
the ve.ui.CXTranslationAction save and restores the scroll position.

So, make sure the scroll position is not changed in placeholder clicks
for translation.

Since ve.ui.CXTranslationAction calls the same target.setSectionContent
method, the redundant scroll position restoration can be removed from there.
But left for a follow up commit and testing.

There are bug reports saying the scrolling also happens in keypress in
translation. This patch does not address that.

Bug: T245855
Change-Id: I8a95976ccd7e302e2888ca4695d6129ab758f05c