SX translator api client: Add getSectionContents method
Publishing a section in a specific position requires fetching
contents of the current section in this position first.

To support this add getSectionContents action inside translator
api client. This method sends a parse request to the action api
and returns a promise that resolves to parser's output as an HTML
string, or null if request was unsuccessful.

Change-Id: I0b1240673f2ce76ebefa67429f1b9f717a8d3e56