SX publishing: Position new section before "References"
New section should be published before "References" and similar
sections inside target article. Since there is no out of the box
support for adding a new section to any valid position inside
an article, we'll need to fetch the contents of the current section
in the desired position and prepend the new section contents to them.

To support publishing new section before "References" and similar
sections:

* Move section number calculation inside publishHelper and use
  appendix titles for target language to correctly calculate it
* Move calculation of html to be published inside publishHelper
* Use these helpers inside translator/publishTranslation action
  translator api client publishTranslation method
* Pass all required publish parameters (page titles, section titles
  languages etc) to translator/publishTranslation api client method
  to handle publishing request

Bug: T263139
Change-Id: I1eec5ab721b76df893d50dc2fe63427fab4887a4