SX Publishing: Add publishHelper to support section concatenation
In order to position a new section in a desired position inside an
article, new section contents need to be concatenated with the
contents of the current section in this position. To support this
need, create publishHelper module that exposes
prependNewSectionToAppendixSection method. Given two PageSection
model instances, this method will return an HTML string containing
the contents of both sections.

Also add tests.

Fix minor typo inside subSection model, too

Change-Id: I822f071b855239fda07db199e7faf107229a4f23