UI lib: Fix MWInput to accept v-model
Until now when input value is updated inside a MWInput component, an
"update" event is emitted with the new value as payload. While this
works, it's doesn't support the usage of "v-model", and developers
should use the "update" event to handle any change in value.

Fix this by emitting "input" event with the proper payload, so that
v-model can be used on MwInput properly. While at it, also add
indicator clicked event emission.

Add stories that demonstrate these enhancements, too.

Change-Id: Ie9a48a374de9f3308fbeedb1b0aec7bfc6215068