SX: Fix JS error on source language update
When updating source language to a language that has not
used before for that session, an error is being thrown,
because of undefined appendixSourceTitles variable inside
fetchSectionSuggestionsBySeeds action.

Fix this by moving fetching of appendix section titles inside
suggestions/actions module and using appendix section titles in
target language that are available.

Screencasts: https://imgur.com/a/qWP5REI

Bug: T283077
Change-Id: I3e117eac8b2ff6f6f307bb98a1e2b8072296548e