SX: Add support for publishing lead section translation
Publishing a translation of a lead section (which leads to the
creation of a new article) is not supported until now. To make
that possible some tweaks around publishing functionality are
needed.

To support lead section translation publishing, add proper
calculation of section number and target article title. Fix
tests accordingly, too.

Bug: T260144
Change-Id: Ic8f381ce3b4fa26af3215096e1f3f4a562f7ef71