Add entrypoint to CX from mobile frontend language searcher
When no languages are found from the search, CX tries to see if the
query string corresponds to a language. And then if that language is
one of the wiki where CX can be used in mobiles, provide an entry
point to CX with language pairs prefilled.

New configuration:
 wgContentTranslationMFLanguageSearchEntrypointTargetLanguages

This has only 'bn' as the value right now.

Bug: T286645
Change-Id: Ib4d78893094953002c26c4b0399fbad1ef2f3082