SX Remove LanguageSearchEntrypointTargetLanguages config parameter
Inside extension.json there exist two configuration parameters that
are very similar:

1. "SectionTranslationTargetLanguages" that refers to the available
   languages that can be selected from the language selector as target
   language.
2. "ContentTranslationMFLanguageSearchEntrypointTargetLanguages" that
   refers to the languages that are enabled for the newly created
   entrypoint through language search in Mobilefrontend.

Since their usage is basically the same, we can (and should) remove
one of them for better clarity and to avoid duplicate variables.

To fix this, remove the second parameter and replace all its
occurrences with "SectionTranslationTargetLanguages" instead.

Bug: T290302
Change-Id: I8e82286a6aa3f81ca0e434210633e4309fc3398c