SXTranslationConfirmer: Enable bookmark action
The bookmark icon exists inside "Confirm Translation" step, however
it didn't work until now, and was marked as TODO.

Enable bookmarking functionality by adding a click handler that
marks/unmarks current suggestion as favorite. Also make icon to
reflect whether the article is already in the "For later" list or
not, both in shape and color.

Bug: T290482
Change-Id: Ie46966e97a66a0d72e3c79ce872c45d1fe4235e5