SXContentComparator: Fix regression issues in useListSelector
useListSelector composable needs access to i18n method so that it can
properly create list messages. Last implementation passes
context.root.$i18n as an argument to useListSelector, however this
doesn't work because the function is invoked at the useListSelector
scope and doesn't have access to "this" keyword, and that leads to
an exception.

In order to fix this issue we could either use "bind" method to bind
context.root, or simply pass context.root to the useListSelector
composable.

Additionally useListSelector also needs the SFCs props in order to
properly work, so we are passing them as argument too.

Change-Id: I719df6d7450dba5b22c73b993454add28a73725c