SX: Add redirection to target wiki on target language change
In order to properly support multiple target languages for production
wikis, the application should redirect to the corresponding wiki when
user updates the target language.

Additionally, when source language changes and the newly selected
source language is the same as the current target language, the
source and target languages should swap and the application redirects
to the wiki of the previously selected source language.

Bug: T291552
Change-Id: Iee0bd38edc54d7482784bcee88574915f81553b1