SX: Add wgSectionTranslationTargetLanguages config as state variable
wgSectionTranslationTargetLanguages configuration parameter is used
in multiple places across the application.

Move this parameter to a vuex mediawiki state variable so that we
have a single source of truth for this piece of information. This
will enable us to easily update the logic in a single place if we
decide to change this implementation.

Change-Id: I906b470c6141a7c1d8a1f17393cdd0b6f5830b06