SX getSubSectionNodes: Fix issue when references are unresolved
Currently, getSubSectionNodes method returns an empty array when
resolveReferences argument if false (i.e. when page is target page).
This issue occurs because jQuery "find" method is used instead of
"filter" (subSectionNodeList is a list of DOM nodes, not a single
node).

Fix this issue by properly filtering html content for CX sections
and then return the list as is.

Change-Id: Iea85c15302fdc4686c960b711c643dc06a9c8bf7