SX Publishing: Fix issues for published target section title
Currently, the target section titles that are being published are not
always valid. More specifically, when creating a new page by
translating the lead section, this new section appears with a title,
and that title being the summary of the edit action. Additionally,
when a missing section is being published and no appendix sections
exist for this article (in this case this section is positioned as
"new", at the bottom of the page), the title of the section is
replaced by the summary of the edit.

This is an expected behaviour, as per the Action API documentation
(https://en.wikipedia.org/w/api.php?action=help&modules=edit), when
when section=new and sectiontitle is not set, summary becomes the
section title.

To fix these issues this patch:
* Fixes getPage getter to return null when empty title provided. This
  bug leads getCurrentTargetPage getter to return a random page,
  instead of null when translating a lead section, messing up the
  calculation of the target section title.
* Move calculateNewSectionNumber and calculateHtmlToPublish methods
  from publishHelper to translator/getters, since these methods
  heavily depend on multiple state properties and other getters.
* Add tests for these getters too.
* Fix bug inside Page/getSectionNumberByTitle method and add tests
  for it.

Bug: T292138
Change-Id: I6b4227a911deaeb5a142c86e3cb342e01fda254a