SX: Move initializeDashboardContext vuex action to composable
application/initializeDashboardContext vuex action contains a lot of
code that is not related to the store itself. Having no store
mutations itself, it is basically a wrapper around other actions.
Some of this actions such "initializeLanguages" and
"loadSectionSuggestionFromUrl" contain a lot of functionality not
related to the store, such as URL parameter handling and logic
for language pair initialization.

To have a better separate of concerns and not bloat the store
actions, move these actions to composables. Also move the related
tests.

Bug: T284058
Change-Id: Ibfc3c2db8f84796ae4c6faee5bc8088c05827b4d